Devops: Add basic Python package scaffolding
The package setup is fully defined by the `pyproject.toml` file. The
code is linted and automatically formatte using `pre-commit` which is
configured in `.pre-commit-config.yaml`.

A Github Action workflow is added that will automatically run
`pre-commit` on push and pull request events as well as run the test
suite using `pytest`.